third_party/WebKit/LayoutTests/http/tests/performance-timing/paint-timing/first-contentful-paint.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FP followed by FCP</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <div id="main"></div>
    <div id="image"></div>

    <script>
      async_test(function (t) {
        testRunner.capturePixelsAsyncThen(t.step_func(function() {
            const bufferedEntries = performance.getEntriesByType('paint');
            assert_equals(bufferedEntries.length, 0, "No paint entries yet");

            const div = document.createElement("div");
            div.style.width = "100px";
            div.style.height = "100px";
            div.style.backgroundColor = "red";
            div.style.color = "blue";
            document.getElementById("main").appendChild(div);

            testRunner.capturePixelsAsyncThen(t.step_func(function() {
                const bufferedEntries = performance.getEntriesByType('paint');
                assert_equals(bufferedEntries.length, 1, "FP only.");
                assert_equals(bufferedEntries[0].entryType, "paint");
                assert_equals(bufferedEntries[0].name, "first-paint");

                const img = document.createElement("IMG");
                img.src = "resources/circles.png";
                img.onload = function() {
                    testRunner.capturePixelsAsyncThen(t.step_func_done(function() {
                        const bufferedEntries = performance.getEntriesByType('paint');
                        assert_equals(bufferedEntries.length, 2, "FP and FCP.");
                        assert_equals(bufferedEntries[0].entryType, "paint");
                        assert_equals(bufferedEntries[0].name, "first-paint");
                        assert_equals(bufferedEntries[1].entryType, "paint");
                        assert_equals(bufferedEntries[1].name, "first-contentful-paint");
                    }));
                };
                document.getElementById('image').appendChild(img);
            }));
        }));
      }, "First Paint triggered by non-contentful paint. Image load triggers First Contentful Paint.");
    </script>
</body>
</html>
